Prevent 'stuck' MIDI notes
Sometimes the MIDI notes get 'stuck', which seems to have been because
the code couldn't handle interlaced KeyPresses and KeyReleases properly.

Having explicit state transitions for the note statuses seems to have
solved this issue for me. It should also make it easier if someone
wishes to implement (*Key)->status = SUSTAINED or similar in the future.

This also made the previous 'key repeat' functionality excessive.